+++ src/capturer/shared.rs
@@ -14,6 +14,7 @@ use windows::{
   },
 };
 
+/// This is not clone-able.
 #[derive(Debug)]
 pub struct SharedMemory {
   buffer: *mut u8,
@@ -45,9 +46,9 @@ impl Drop for SharedMemory {
 }
 
 /// Capture screen to a chunk of shared memory.
-pub type SharedCapturer = Capturer<SharedMemory>;
+pub type SharedMemoryCapturer = Capturer<SharedMemory>;
 
-impl SharedCapturer {
+impl SharedMemoryCapturer {
   pub fn create(monitor: Monitor, name: &str) -> Result<Self> {
     let (buffer, buffer_size, file, texture, texture_desc) = Self::allocate(&monitor, name)?;
     Ok(Self {
@@ -168,7 +169,7 @@ mod tests {
   #[serial]
   fn shared_capturer() {
     let monitor = Scanner::new().unwrap().next().unwrap();
-    let mut capturer = SharedCapturer::create(monitor, "RustyDuplicationTest").unwrap();
+    let mut capturer = SharedMemoryCapturer::create(monitor, "RustyDuplicationTest").unwrap();
 
     // sleep for a while before capture to wait system to update the screen
     thread::sleep(Duration::from_millis(100));